checkpoint - cursor doesn't read/write in parallel well
